requirements.txt
streamlit
pydantic
requests
langchain
langchain-anthropic
langchain-core
langchain-community
python-dotenv